make settings form respect existing settings
Also switches from on/off Sharing enabled to selecting which methods are available for sharing. There's probably a UI/UX experience question there, and how to handle site-module settings that are, or could be, more than on/off.

Not setting anything produces a validation error, which I think is a related issue for discussion.